Move geo functions out of SoundscapeService and into their own class
There's nothing clever going on here, mostly just moving functions around.
The only other changes are:

 1. The locale configuration is created and held in the GeoEngine to save
  re-initializing it each time it's required,
 2. All functions that require the application context are made within
  the start() function so that it's only required in that one place.
 3. The myLocation, whatsAroundMe and aheadOfMe functions now all return
  lists of Strings (or PositionedStrings) so that the text can be used
  in the UI and not just the speech.
